Zephyr: Added in our preliminary support for touch on the display
Zephyr: Added in our preliminary support for touch on the display

In ArduinoCore-zephyr added a callback to register with the
input class.  When set it simply forwards the callbacks to
this function.

Then added some support.  We still don't have things like
gestures as I don't think the zephyr input class has support
for it.

Have a simple touch paint sketch (now supporting multiple
touches, that appears to work

Edit: The header file now combines the two different
versions of the class into one.

Added #ifdef __MBED__ in the mbed implementation file so that
the code would not compile when we are doing zephyr builds